+++ shareaza/DownloadTransferHTTP.cpp
@@ -1302,7 +1302,7 @@ BOOL CDownloadTransferHTTP::ReadContent()
 	{
 		m_pSource->SetValid();
 
-		size_t nLength	= min( pInput->m_nLength, m_nLength - m_nPosition );
+		QWORD nLength	= min( (QWORD)pInput->m_nLength, m_nLength - m_nPosition );
 		BOOL bSubmit	= FALSE;
 
 		if ( m_bChunked )
@@ -1369,7 +1369,7 @@ BOOL CDownloadTransferHTTP::ReadContent()
 				{
 					// m_nChunkLength must be smaller than nLength so this cast
 					// is safe
-					nLength = static_cast< size_t >( m_nChunkLength );
+					nLength = m_nChunkLength;
 				}
 				m_nChunkLength -= nLength;
 				if ( m_nChunkLength == 0 )
@@ -1592,7 +1592,7 @@ BOOL CDownloadTransferHTTP::ReadFlush()
 
 	if ( m_nContentLength == SIZE_UNKNOWN ) m_nContentLength = 0;
 	
-	DWORD nRemove = (DWORD)min( pInput->m_nLength, m_nContentLength );
+	QWORD nRemove = min( (QWORD)pInput->m_nLength, m_nContentLength );
 	m_nContentLength -= nRemove;
 	
 	pInput->Remove( nRemove );
