+++ shareaza/LibraryBuilder.cpp
@@ -499,7 +499,7 @@ bool CLibraryBuilder::HashFile(LPCTSTR szPath, HANDLE hFile)
 	QWORD nLength = nFileSize;
 	while ( nLength )
 	{
-		nBlock	= (DWORD)min( nLength, MAX_HASH_BUFFER_SIZE );
+		nBlock	= (DWORD)min( nLength, (QWORD)MAX_HASH_BUFFER_SIZE );
 
 		{
 			// Calculate % done (nResult = 0 -> 100)
